--- examples/ann_brain_with_feedback/README.md
@@ -1,5 +1,0 @@
-This example shows you how to control a modular robot using a simple feed-forward artificial neural network and feedback from servo positions.
-
-You learn:
-- How to create and read sensors
-- How to control a modular robot using a simple feed-forward artificial neural network.
