+++ examples/brain_with_feedback/README.md
@@ -1,0 +1,5 @@
+This example shows you how to get feedback in the brain from the current active hinge positions.
+
+You learn:
+- How to create and read sensors
+- How to control read sensors in the brain.
