--- todo
@@ -1,2 +1,0 @@
-update docs
-Make a good controller for the example.
